Extract weights update group lifecycle to free functions in weight_updater
